+++ vms/evm/sync/hashdb/BUILD.bazel
@@ -57,3 +57,60 @@ go_test(
         "@com_github_stretchr_testify//require",
     ],
 )
+
+go_library(
+    name = "hashdb",
+    srcs = [
+        "client.go",
+        "handler.go",
+        "snapshot.go",
+    ],
+    importpath = "github.com/ava-labs/avalanchego/vms/evm/sync/hashdb",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//ids",
+        "//network/p2p",
+        "//proto/pb/sync",
+        "//snow/engine/common",
+        "//utils/logging",
+        "//vms/evm/sync/handlers",
+        "//vms/evm/sync/network",
+        "@com_github_ava_labs_libevm//common",
+        "@com_github_ava_labs_libevm//core/rawdb",
+        "@com_github_ava_labs_libevm//core/state/snapshot",
+        "@com_github_ava_labs_libevm//core/types",
+        "@com_github_ava_labs_libevm//crypto",
+        "@com_github_ava_labs_libevm//ethdb",
+        "@com_github_ava_labs_libevm//ethdb/memorydb",
+        "@com_github_ava_labs_libevm//libevm/options",
+        "@com_github_ava_labs_libevm//trie",
+        "@com_github_ava_labs_libevm//triedb",
+        "@org_uber_go_zap//:zap",
+    ],
+)
+
+go_test(
+    name = "hashdb_test",
+    srcs = [
+        "client_test.go",
+        "handler_test.go",
+    ],
+    embed = [":hashdb"],
+    deps = [
+        "//ids",
+        "//network/p2p",
+        "//proto/pb/sync",
+        "//snow/engine/common",
+        "//utils",
+        "//utils/logging",
+        "//utils/logging/loggingtest",
+        "//vms/evm/sync/synctest",
+        "@com_github_ava_labs_libevm//common",
+        "@com_github_ava_labs_libevm//core/types",
+        "@com_github_ava_labs_libevm//trie",
+        "@com_github_ava_labs_libevm//triedb",
+        "@com_github_google_go_cmp//cmp",
+        "@com_github_google_go_cmp//cmp/cmpopts",
+        "@com_github_stretchr_testify//require",
+    ],
+)
